+++ comics/comics/peanuts.py
@@ -1,15 +1,15 @@
-from comics.aggregator.crawler import ComicsComCrawlerBase
+from comics.aggregator.crawler import GoComicsComCrawlerBase
 from comics.meta.base import MetaBase
 
 class Meta(MetaBase):
     name = 'Peanuts'
     language = 'en'
-    url = 'http://comics.com/peanuts/'
+    url = 'http://www.gocomics.com/peanuts/'
     start_date = '1950-10-02'
     end_date = '2000-02-13'
     rights = 'Charles M. Schulz'
 
-class Crawler(ComicsComCrawlerBase):
+class Crawler(GoComicsComCrawlerBase):
     history_capable_date = '1950-10-02'
     schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'